Revert "al - card expansion"
