+++ Roomi/Roomi/Core/Authentication/View/ProfileCreationView.swift
@@ -26,7 +26,6 @@ struct ProfileCreationView: View, Hashable {
             
             // name input
             InputView(text: $viewModel.potentialUser.name, title: "Name", placeholder: "Name")
-                .textInputAutocapitalization(.words)
             
             // age selector
             Text("Age")
