+++ Roomi/Roomi/Core/Root/ContentView.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 struct ContentView: View {
     @EnvironmentObject var viewModel: AuthViewModel
-    @EnvironmentObject var relationshipsViewModel: UserRelationshipsViewModel
     
     var body: some View {
         Group {
@@ -31,5 +30,4 @@ struct ContentView: View {
 #Preview {
     ContentView()
         .environmentObject(AuthViewModel())
-        .environmentObject(UserRelationshipsViewModel())
 }
